requirements.txt
pywikibot
python-dotenv
numpy
matplotlib